requirements.txt
gunicorn
psycopg2
django-environ
djangorestframework
djoser
django-cors-headers